cont : refactor
