+++ common/arg.cpp
@@ -2427,16 +2427,16 @@ common_params_context common_params_parser_init(common_params & params, llama_ex
         }
     ).set_env("LLAMA_ARG_FIT"));
     add_opt(common_arg(
-        { "-fite", "--fit-estimate" }, "[on|off]",
-        string_format("estimate the required memory to run the model ('on' or 'off', default: '%s')", params.fit_params_est ? "on" : "off"),
+        { "-fitp", "--fit-print" }, "[on|off]",
+        string_format("print the estimated required memory ('on' or 'off', default: '%s')", params.fit_params_print ? "on" : "off"),
         [](common_params & params, const std::string & value) {
             if (is_truthy(value)) {
-                params.fit_params_est = true;
+                params.fit_params_print = true;
             } else if (is_falsey(value)) {
-                params.fit_params_est = false;
+                params.fit_params_print = false;
             } else {
                 throw std::runtime_error(
-                    string_format("error: unknown value for --fit-estimate: '%s'\n", value.c_str()));
+                    string_format("error: unknown value for --fit-print: '%s'\n", value.c_str()));
             }
         }
     ).set_examples({LLAMA_EXAMPLE_FIT_PARAMS}).set_env("LLAMA_ARG_FIT_ESTIMATE"));
