+++ tools/fit-params/fit-params.cpp
@@ -3,6 +3,7 @@
 
 #include "arg.h"
 #include "common.h"
+#include "fit.h"
 #include "log.h"
 
 #include <cinttypes>
@@ -26,8 +27,8 @@ int main(int argc, char ** argv) {
     auto mparams = common_model_params_to_llama(params);
     auto cparams = common_context_params_to_llama(params);
 
-    if (!params.fit_params_est) {
-        const llama_params_fit_status status = llama_params_fit(params.model.path.c_str(), &mparams, &cparams,
+    if (!params.fit_params_print) {
+        const common_params_fit_status status = common_params_fit(params.model.path.c_str(), &mparams, &cparams,
                 params.tensor_split, params.tensor_buft_overrides.data(), params.fit_params_target.data(), params.fit_params_min_ctx,
                 params.verbosity >= 4 ? GGML_LOG_LEVEL_DEBUG : GGML_LOG_LEVEL_ERROR);
         if (status != LLAMA_PARAMS_FIT_STATUS_SUCCESS) {
